ui: f32 to display not integer number

+++ src/ui.rs
@@ -43,7 +43,10 @@ fn update_health(
     for hp in hp_query.iter() {
         let mut text = text_query.single_mut();
         text.sections[0] = TextSection::new(
-            format!("Health: {}%", (hp.current / hp.max) * 100),
+            format!(
+                "Health: {}%",
+                ((hp.current as f32) / (hp.max as f32)) * 100.
+            ),
             text.sections[0].style.clone(),
         );
     }
